+++ R/get_vpts.R
@@ -63,6 +63,7 @@ get_vpts <- function(radar,
                      datetime,
                      source = c("baltrad", "uva", "ecog-04003", "rmi"),
                      return_type = c("vpts", "tibble")) {
+  # Input checks ----
   # Check source argument
   ## If no source is provided, set "baltrad" as default
   if (missing(source)) {
@@ -130,7 +131,7 @@ get_vpts <- function(radar,
       class = "getRad_error_date_parsable"
     )
   }
-  # Parse the provided date argument to a lubridate interval
+  # Parse the provided date argument to a lubridate interval ----
   ## If the date is a single date, convert it to an interval
   if (!inherits(datetime, "Interval")) {
     datetime_converted <- lubridate::as_datetime(datetime)
@@ -173,16 +174,15 @@ get_vpts <- function(radar,
     date_interval <- datetime
   }
 
-  ## We need to round the interval because the helpers always fetch data a day
-  ## at a time
+  ## Round the interval because the helpers always fetch data a day at a time ----
   date_interval_utc <- lubridate::as.interval(
     lubridate::with_tz(lubridate::int_start(date_interval), "UTC"),
-    lubridate::with_tz(lubridate::int_start(date_interval), "UTC")
+    lubridate::with_tz(lubridate::int_end(date_interval), "UTC")
   )
   rounded_interval <- round_interval(date_interval_utc, "day")
 
-  # Query the selected radars by directing to the correct get_vpts_* helper
-  # based on source.
+  # Query the selected radars ----
+  # Directing to the correct get_vpts_* helper based on source.
   cl <- rlang::caller_env(0)
   fetched_vpts <-
     switch(dplyr::case_when(
@@ -198,7 +198,7 @@ get_vpts <- function(radar,
     ) |> radar_to_name()
 
 
-  # Drop any results outside the requested interval
+  # Drop any results outside the requested interval ----
   filtered_vpts <-
     fetched_vpts |>
     purrr::map(
@@ -218,7 +218,7 @@ get_vpts <- function(radar,
       },
       .purrr_error_call = cl
     )
-  # Return the vpts data
+  # Return the vpts data ----
   ## By default, return drop the source column and convert to a vpts object for
   ## usage in bioRAD
   return_type <- rlang::arg_match(return_type)
@@ -239,6 +239,6 @@ get_vpts <- function(radar,
         }
       })(filtered_vpts)
     )
-  # Return the converted/formatted object
+  # Return the converted/formatted object ----
   return(return_object)
 }