+++ tests/testthat/test-get_pvol_ro.R
@@ -1,0 +1,30 @@
+test_that("Check if the available attributes changed", {
+  skip_if_offline()
+  expect_identical(
+    httr2::request("https://opendata.meteoromania.ro/radar/MED/") |>
+      httr2::req_perform() |>
+      httr2::resp_body_html() |>
+      xml2::xml_find_all("//a/@href") |>
+      xml2::xml_text() |> tail(-1) |>
+      gsub(pattern = ".*0200", replacement = "") |>
+      unique() |> gsub(pattern = ".hdf", replacement = ""),
+    c("KDP", "RhoHV", "V", "ZDR", "dBR", "dBZ", "Height")
+  )
+})
+test_that("Pvol for Romania can be downloaded", {
+  skip_if_offline()
+  time <- lubridate::floor_date(
+    as.POSIXct(Sys.time(),
+      tz = "Europe/Helsinki"
+    ) - lubridate::hours(10), "5 mins"
+  )
+  pvol <- expect_s3_class(get_pvol("romed",
+    time,
+    param = "all"
+  ), "pvol")
+  expect_true(bioRad::is.pvol(pvol))
+  expect_identical(
+    lubridate::floor_date(pvol$datetime, "5 mins"),
+    lubridate::with_tz(time, "UTC")
+  )
+})